move scripts and tools to scripts/ folder and go stuff to cmd/
