--- apps/cli/main.go
@@ -1,7 +1,0 @@
-package main
-
-import "github.com/skewer-project/skewer/apps/cli/cmd"
-
-func main() {
-	cmd.Execute()
-}
