+++ cmd/cli/main.go
@@ -1,0 +1,5 @@
+package cli
+
+func main() {
+	Execute()
+}
