+++ cmd/coordinator/server_path_test.go
